fixing scripts to also copy fixed traces to traces
