+++ fase23clean.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 # Cleaning the working folders for the FASE-23 experiment
 echo "** Cleaning generated test suites..."
-rm -r ../MCtest/*
-rm -r ../SBTtest/*
+rm -rf ../MCtest/*
+rm -rf ../SBTtest/*
 echo "** Cleaning the Combinedtest-dir..."
-rm -r ../Combinedtest/*.txt
-rm -r ../Combinedtest/*.ser
-rm -r ../Combinedtest/Model/*
+rm -rf ../Combinedtest/*.txt
+rm -rf ../Combinedtest/*.ser
+rm -rf ../Combinedtest/Model/*
 echo "** Cleaning the traces-dir..."
-rm ../traces/*
+rm -rf ../traces/*
 echo "** Done"
